feat(BeitV2): modify some format for review
